Refactors booking request into service class
