+++ tests/php/Model/BookingTest.php
@@ -9,6 +9,7 @@ use CommonsBooking\Model\Location;
 use CommonsBooking\Model\Restriction;
 use CommonsBooking\Model\Timeframe;
 use CommonsBooking\Plugin;
+use CommonsBooking\Service\Booking as BookingAlias;
 use CommonsBooking\Tests\Wordpress\CustomPostTypeTest;
 use SlopeIt\ClockMock\ClockMock;
 
@@ -745,7 +746,7 @@ class BookingTest extends CustomPostTypeTest {
 		$endingTime->setTime( 17, 59, 59 );
 		// we need to create this booking in the "frontend" way in order to save the correct grid sizes for the generation
 		// pickup and returntimes
-		$testBookingSpanningOverTwoSlotsID     = \CommonsBooking\Wordpress\CustomPostType\Booking::handleBookingRequest(
+		$testBookingSpanningOverTwoSlotsID     = BookingAlias::handleBookingRequest(
 			$separateItem,
 			$separateLocation,
 			'confirmed',
