+++ tests/php/Wordpress/CustomPostType/BookingTest.php
@@ -2,6 +2,7 @@
 
 namespace CommonsBooking\Tests\Wordpress\CustomPostType;
 
+use CommonsBooking\Service\Booking as BookingAlias;
 use CommonsBooking\Tests\Wordpress\CustomPostTypeTest;
 use CommonsBooking\Wordpress\CustomPostType\Booking;
 use SlopeIt\ClockMock\ClockMock;
@@ -27,7 +28,7 @@ class BookingTest extends CustomPostTypeTest {
 		$date->modify( '-1 day' );
 		ClockMock::freeze( $date );
 		// Case 1: We create an unconfirmed booking for a bookable timeframe. The unconfirmed booking should be created
-		$bookingId = Booking::handleBookingRequest(
+		$bookingId = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -50,7 +51,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->assertFalse( $bookingModel->isConfirmed() );
 
 		// Case 2: We now confirm the booking. The booking should be confirmed
-		$newBookingId       = Booking::handleBookingRequest(
+		$newBookingId       = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
@@ -73,7 +74,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 3: We now try to cancel our booking a little bit later. The booking should be cancelled.
 		$date->modify( '+ 5 hours' );
 		ClockMock::freeze( $date );
-		$canceledId         = Booking::handleBookingRequest(
+		$canceledId         = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'canceled',
@@ -100,7 +101,7 @@ class BookingTest extends CustomPostTypeTest {
 	public function testHandleBookingRequest_deleteUnconfirmed() {
 		ClockMock::freeze( new \DateTime( self::CURRENT_DATE ) );
 		// We create an unconfirmed booking and then cancel the booking. The booking should be canceled
-		$bookingId          = Booking::handleBookingRequest(
+		$bookingId          = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -119,7 +120,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->assertTrue( $bookingModel->isUnconfirmed() );
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Booking canceled.' );
-		Booking::handleBookingRequest(
+		BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'delete_unconfirmed',
@@ -139,7 +140,7 @@ class BookingTest extends CustomPostTypeTest {
 		$date->modify( '-1 day' );
 		ClockMock::freeze( $date );
 		// 3 Days are overbooked, that means that the Litepicker had 3 locked / holidays in range
-		$bookingId = Booking::handleBookingRequest(
+		$bookingId = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -163,7 +164,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->assertFalse( $bookingModel->isConfirmed() );
 
 		// The overbooked days are not present anymore when confirming the booking cause they are only calculated on the Litepicker screen
-		$newBookingId       = Booking::handleBookingRequest(
+		$newBookingId       = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
@@ -190,7 +191,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 1: We try to create a booking without a defined location
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Location does not exist. ()' );
-		$booking = Booking::handleBookingRequest(
+		$booking = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			null,
 			'unconfirmed',
@@ -206,7 +207,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 2: We try to create a booking without a defined item
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Item does not exist. ()' );
-		$booking = Booking::handleBookingRequest(
+		$booking = BookingAlias::handleBookingRequest(
 			null,
 			$this->locationId,
 			'unconfirmed',
@@ -222,7 +223,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 3: We try to create a booking without a defined start date
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Start- and/or end-date is missing.' );
-		$booking = Booking::handleBookingRequest(
+		$booking = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -238,7 +239,7 @@ class BookingTest extends CustomPostTypeTest {
 		// Case 4: We try to create a booking without a defined end date
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'Start- and/or end-date is missing.' );
-		$booking = Booking::handleBookingRequest(
+		$booking = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -255,7 +256,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->createConfirmedBookingStartingToday();
 		$this->expectException( \CommonsBooking\Exception\BookingDeniedException::class );
 		$this->expectExceptionMessage( 'There is already a booking in this time-range. This notice may also appear if there is an unconfirmed booking in the requested period. Unconfirmed bookings are deleted after about 10 minutes. Please try again in a few minutes.' );
-		$booking = Booking::handleBookingRequest(
+		$booking = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -270,7 +271,7 @@ class BookingTest extends CustomPostTypeTest {
 
 	public function testReAccessUnconfirmed() {
 		// this tests the case where the same user tries to access their unconfirmed booking again
-		$bookingId = Booking::handleBookingRequest(
+		$bookingId = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -287,7 +288,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->assertIsInt( $bookingId );
 		$bookingModel = new \CommonsBooking\Model\Booking( $bookingId );
 		$this->assertTrue( $bookingModel->isUnconfirmed() );
-		$sameBookingId      = Booking::handleBookingRequest(
+		$sameBookingId      = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -314,7 +315,7 @@ class BookingTest extends CustomPostTypeTest {
 	 * @return void
 	 */
 	public function testHandleBookingRequest_deleted_confirm() {
-		$bookingId = Booking::handleBookingRequest(
+		$bookingId = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'unconfirmed',
@@ -334,7 +335,7 @@ class BookingTest extends CustomPostTypeTest {
 		$this->expectExceptionMessage( 'Your reservation has expired, please try to book again' );
 
 		// now we try to confirm the booking
-		$confirmedId = Booking::handleBookingRequest(
+		$confirmedId = BookingAlias::handleBookingRequest(
 			$this->itemId,
 			$this->locationId,
 			'confirmed',
